make cached preparedQuery a generic feature in vertx-sql-client and add TCK tests for it
