+++ vertx-pg-client/src/test/java/io/vertx/pgclient/PreparedStatementCachedTest.java
@@ -17,38 +17,11 @@
 
 package io.vertx.pgclient;
 
-import io.vertx.sqlclient.Tuple;
-import io.vertx.ext.unit.Async;
-import io.vertx.ext.unit.TestContext;
-import org.junit.Test;
-
 public class PreparedStatementCachedTest extends PreparedStatementTestBase {
 
   @Override
   protected PgConnectOptions options() {
     return new PgConnectOptions(options).setCachePreparedStatements(true);
-  }
-
-  @Test
-  public void testConcurrent(TestContext ctx) {
-    PgConnection.connect(vertx, options(), ctx.asyncAssertSuccess(conn -> {
-      Async[] asyncs = new Async[10];
-      for (int i = 0;i < 10;i++) {
-        asyncs[i] = ctx.async();
-      }
-      for (int i = 0;i < 10;i++) {
-        Async async = asyncs[i];
-        conn.prepare("SELECT * FROM Fortune WHERE id=$1", ctx.asyncAssertSuccess(ps -> {
-          ps.execute(Tuple.of(1), ctx.asyncAssertSuccess(results -> {
-            ctx.assertEquals(1, results.size());
-            Tuple row = results.iterator().next();
-            ctx.assertEquals(1, row.getInteger(0));
-            ctx.assertEquals("fortune: No such file or directory", row.getString(1));
-            async.complete();
-          }));
-        }));
-      }
-    }));
   }
 
 }
