+++ vertx-pg-client/src/test/java/io/vertx/pgclient/tck/PgPreparedQueryPooledTest.java
@@ -7,6 +7,7 @@ import org.junit.runner.RunWith;
 public class PgPreparedQueryPooledTest extends PgPreparedQueryTestBase {
   @Override
   protected void initConnector() {
-    connector = ClientConfig.POOLED.connect(vertx, rule.options());
+    options = rule.options();
+    connector = ClientConfig.POOLED.connect(vertx, options);
   }
 }
